states

+++ src/1_systems/t_tick_system.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
+#include "0_engine/g_globals.hpp"
 #include "0_engine/u_ecs.hpp"
 #include "0_engine/u_types.hpp"
-#include "1_systems/u_orchestra.hpp"
 
 namespace pce {
 using Tick = StrongType<u32, struct TickTag, Arithmetic>;
@@ -16,7 +16,7 @@ struct TickSystem {
     void operator()() const {
         TickState& tick_state = Singleton::Get<TickState>();
         const nanoseconds64 time { TimeNowNS() };
-        tick_state.delta_time = (time - tick_state.last_tick_time).value * NS_TO_SECONDS;
+        tick_state.delta_time = static_cast<f32>((time - tick_state.last_tick_time).value) * NS_TO_SECONDS;
         tick_state.last_tick_time = time;
         tick_state.tick += Tick { 1U };
     }
